% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.preprocess.R
\name{data.preprocess}
\alias{data.preprocess}
\title{data.preprocess}
\usage{
data.preprocess(
  data,
  resp,
  event = NULL,
  trt,
  stratcov = NULL,
  basecov.cont = NULL,
  basecov.cat = NULL
)
}
\arguments{
\item{data}{input dataset (a data frame)}

\item{resp}{the name of the response variable, for survival endpoint, it is the time-to-event variable}

\item{event}{(for survival endpoint only) the name of the event variable, which takes value 1 for event and 0 for censored, or NULL for other types of endpoint}

\item{trt}{the name of the treatment variable, which takes value 1 for experimental treatment and 0 for control treatment}

\item{stratcov}{a vector containing the name(s) of stratification factors, or NULL}

\item{basecov.cont}{a vector containing the name(s) of continuous baseline covariates, or NULL}

\item{basecov.cat}{a vector containing the name(s) of categorical baseline covariates, or NULL}
}
\value{
A matrix with required variables in specific order
}
\description{
{ Preprocess input dataset to include required variables and output as a matrix }
}
